--- src/mocks/handlers/board/resolver/get/index.ts
@@ -1,3 +1,0 @@
-export { default as MovieLogReadResolver } from "./all-movie-log";
-export { default as UserWriterMovieLogReadResolver } from "./user-movie-log";
-export { default as SeletorMovieLogReadResolver } from "./selector-movie";
